server/requirements.txt
fastapi
uvicorn
pandas
numpy
pydantic
openai
requests
python-multipart
pillow